Extract StormTopologyBuilderTests
